Move components out of /private and /public
